[CHORE] #60 : userId type int -> Long

+++ android/core/datastore/src/main/java/com/jaino/datastore/BeJuRyuDatastore.kt
@@ -4,7 +4,7 @@ interface BeJuRyuDatastore {
     var accessToken: String
     var refreshToken: String
     var nickName : String
-    var userId: Int
+    var userId: Long
 
     fun clear()
 }
